+++ packages/plugin-packages/mashroom-portal/src/backend/utils/logging_utils.ts
@@ -1,9 +1,13 @@
-
 import type {MashroomPortalApp} from '../../../type-definitions';
 
 export const portalAppContext = (portalApp: MashroomPortalApp): any => {
+    let portalAppHost = null;
+    if (portalApp.remoteApp && portalApp.resourcesRootUri.indexOf('://') !== -1) {
+        portalAppHost = portalApp.resourcesRootUri.split('/')[2];
+    }
     return {
-      portalAppName: portalApp.name,
-      portalAppVersion: portalApp.version
+        portalAppName: portalApp.name,
+        portalAppVersion: portalApp.version,
+        portalAppHost,
     };
 };